requirements.txt
markitdown[all]
openpyxl
pandas
tabulate
openai
python-dotenv
